+++ src/main/java/com/mango/amango/domain/product/presentation/dto/request/CreateProductReq.java
@@ -1,10 +1,9 @@
-package com.mango.amango.domain.product.entity.dto.request;
+package com.mango.amango.domain.product.presentation.dto.request;
 
 import com.mango.amango.domain.tag.entity.Category;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
-import java.time.LocalDateTime;
 import java.util.List;
 
 public record CreateProductReq(
