+++ src/main/java/com/mango/amango/domain/product/service/CreateProductService.java
@@ -1,6 +1,6 @@
 package com.mango.amango.domain.product.service;
 
-import com.mango.amango.domain.product.entity.dto.request.CreateProductReq;
+import com.mango.amango.domain.product.presentation.dto.request.CreateProductReq;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
